requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
PyJWT>=2.8.0
cryptography>=42.0.0
openpyxl>=3.1.0
httpx>=0.27.0
